requirements.txt
streamlit
langchain
langchain-google-genai
langchain-chroma
langgraph
langchain-community
faiss
python-dotenv
unstructured
pypdf
tiktoken
google-generativeai
typing-extensions
pydantic
faiss-cpu